update attune and callAPI
