+++ cogs/tp.py
@@ -4,7 +4,7 @@ import requests
 import re
 from discord.utils import get        
 from discord.ext import commands
-from bfunc import db, headers, commandPrefix, numberEmojis, roleArray
+from bfunc import db, headers, commandPrefix, numberEmojis, roleArray, callAPI, checkForChar
 
 class Tp(commands.Cog):
     def __init__ (self, bot):
@@ -14,6 +14,7 @@ class Tp(commands.Cog):
     async def tp(self, ctx):	
         pass
       
+    #TODO: Stat bonuses for unattuned items
     @tp.command()
     async def buy(self, ctx , charName, mItem):
 
@@ -21,7 +22,7 @@ class Tp(commands.Cog):
         author = ctx.author
         tpEmbed = discord.Embed()
         tpCog = self.bot.get_cog('Tp')
-        charRecords, tpEmbedmsg = await tpCog.checkForChar(ctx, charName, tpEmbed)
+        charRecords, tpEmbedmsg = await checkForChar(ctx, charName, tpEmbed)
 
         if charRecords:
             def tpChoiceEmbedCheck(r, u):
@@ -35,7 +36,7 @@ class Tp(commands.Cog):
                     sameMessage = True
                 return ((str(r.emoji) == '✅') or (str(r.emoji) == '❌')) and u == author
 
-            mRecord = await tpCog.callAPI('MIT',mItem) 
+            mRecord = callAPI('MIT',mItem) 
             if mRecord:
                 if mRecord['Name'] in charRecords['Magic Items']:
                     await channel.send(f"You already have `{mRecord['Name']}`, and cannot spend more TP/GP on another one.")
@@ -200,7 +201,7 @@ class Tp(commands.Cog):
         author = ctx.author
         tpEmbed = discord.Embed()
         tpCog = self.bot.get_cog('Tp')
-        charRecords, tpEmbedmsg = await tpCog.checkForChar(ctx, charName, tpEmbed)
+        charRecords, tpEmbedmsg = await checkForChar(ctx, charName, tpEmbed)
 
         def tpEmbedCheck(r, u):
             sameMessage = False
@@ -261,7 +262,7 @@ class Tp(commands.Cog):
             await channel.send(f"`{tierNum}` is not a valid tier. Please try again with 1,2,3, or 4 or (Junior, Journey, Elite, or True)")
             return
 
-        charRecords, tpEmbedmsg = await tpCog.checkForChar(ctx, charName, tpEmbed)
+        charRecords, tpEmbedmsg = await checkForChar(ctx, charName, tpEmbed)
 
         if charRecords:
             def tpEmbedCheck(r, u):
@@ -314,86 +315,6 @@ class Tp(commands.Cog):
                         tpEmbed.description = f"I have abandoned your T{role} TP!"
                         await tpEmbedmsg.edit(embed=tpEmbed)
 
-        
-    async def checkForChar(self, ctx, char, charEmbed=""):
-        channel = ctx.channel
-        author = ctx.author
-        guild = ctx.guild
-
-        playersCollection = db.players
-        charRecords = list(playersCollection.find({"User ID": str(author.id), "Name": {"$regex": char, '$options': 'i' }}))
-
-        if charRecords == list():
-            await channel.send(content=f'I was not able to find your character named {char}. Please check your spelling and try again')
-            self.char.get_command(ctx.invoked_with).reset_cooldown(ctx)
-            return None, None
-
-        else:
-            if len(charRecords) > 1:
-                infoString = ""
-                charRecords = list(charRecords)
-                print(charRecords)
-                for i in range(0, min(len(charRecords), 9)):
-                    infoString += f"{numberEmojis[i]}: {charRecords[i]['Name']}\n"
-                
-                try:
-                    def infoCharEmbedcheck(r, u):
-                        sameMessage = False
-                        if charEmbedmsg.id == r.message.id:
-                            sameMessage = True
-                        return (r.emoji in numberEmojis[:min(len(charRecords), 9)]) or (str(r.emoji) == '❌') and u == author
-
-                    charEmbed.add_field(name=f"There seems to be multiple results for `{char}`, please choose the correct character.", value=infoString, inline=False)
-                    charEmbedmsg = await channel.send(embed=charEmbed)
-                    for num in range(0,min(len(charRecords), 6)): await charEmbedmsg.add_reaction(numberEmojis[num])
-                    await charEmbedmsg.add_reaction('❌')
-                    tReaction, tUser = await self.bot.wait_for("reaction_add", check=infoCharEmbedcheck, timeout=60)
-                except asyncio.TimeoutError:
-                    await charEmbedmsg.delete()
-                    await channel.send('Character information timed out! Try using the command again')
-                    self.char.get_command(ctx.invoked_with).reset_cooldown(ctx)
-                    return None, None
-                else:
-                    if tReaction.emoji == '❌':
-                        await charEmbedmsg.edit(embed=None, content=f"Character information canceled. User `{commandPrefix}char info` command and try again!")
-                        await charEmbedmsg.clear_reactions()
-                        self.char.get_command(ctx.invoked_with).reset_cooldown(ctx)
-                        return None, None
-                charEmbed.clear_fields()
-                await charEmbedmsg.clear_reactions()
-                return charRecords[int(tReaction.emoji[0]) - 1], charEmbedmsg
-
-        return charRecords[0], None
-
-    async def callAPI(self, table, query):
-        if query == "":
-            return False
-
-        API_URL = ('https://api.airtable.com/v0/appF4hiT6A0ISAhUu/'+ table +'?&filterByFormula=(FIND(LOWER(SUBSTITUTE("' + query.replace(" ", "%20") + '"," ","")),LOWER(SUBSTITUTE({Name}," ",""))))').replace("+", "%2B") 
-        r = requests.get(API_URL, headers=headers)
-        r = r.json()
-
-        if r['records'] == list():
-            return False
-        else:
-            if (len(r['records']) > 1):
-                if table == 'Races' or table == "Background":
-                    for x in r['records']:
-                        print(x['fields']['Name'])
-                        print(query)
-                        if len(x['fields']['Name'].replace(" ", "")) == len(query.replace(" ", "")):
-                            return x['fields']
-
-                if table == 'RIT':
-                    minimum = {'fields': {'Tier': 0}}
-                    for x in r['records']:
-                        if int(x['fields']['Tier']) > int(minimum['fields']['Tier']):
-                            min = x
-                    
-                    return min['fields']
-
-            else:
-                return r['records'][0]['fields']
 
 def setup(bot):
     bot.add_cog(Tp(bot))
